- fixed poi_type bug
git-svn-id: https://gpsdrive.svn.sourceforge.net/svnroot/gpsdrive/trunk@2318 e3ca9d6f-f84a-410d-b97a-e7ddc331358d

+++ src/poi.h
@@ -87,7 +87,7 @@ typedef struct
 
 typedef struct
 {
-	gint id;
+	glong id;
 	GtkListStore *extra_list;
 	GtkWidget *entry_name;
 	GtkWidget *entry_comment;
